+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [5.0.3] - 2021-01-07
 ### Changed
+- mysql_class.SlaveRep.\_\_init\_\_:  Added self.rep_user and self.rep_japd for replication user information.
 - mysql_class.Server.connect:  Set any error connection messages to self.conn_msg.
 - mysql_class.Server.connect:  Added "silent" keyword argument to prevent printing any connection error messages.
-- mysql_class.Server.\_\_init\_\_:  Added conn_msg to handle connection error messages.
+- mysql_class.Server.\_\_init\_\_:  Added self.conn_msg to handle connection error messages.
 - mysql_class.Server.upd_srv_stat:  Replaced percentage equations with calls to gen_libs.pct_int.